project-root/backend/real_adk/requirements.txt
# Google ADK and AI Dependencies
google-adk>=0.2.0
google-generativeai>=0.3.2
google-cloud-documentai>=2.33.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.19.0

# Flask and API
Flask==3.1.2
flask-cors==5.0.0
werkzeug>=3.0.0

# Utilities
python-dotenv==1.0.0
requests>=2.32.0

# Optional: For enhanced functionality
# firebase-admin>=6.6.0  # If using Firebase Auth
# google-cloud-secret-manager>=2.21.1  # For secret management
